+++ src/reducers/SeenPhrasesReducer.js
