+++ src/reducers/index.js
@@ -1,8 +1,8 @@
 import {combineReducers} from 'redux';
 import {categoryList} from './categoryListReducer';
-import {phrases} from './phrasesReducer';
-import {seenPhrases} from './seenPhrasesReducer';
-import {learntPhrases} from './learntPhrasesReducer';
+import {phrases} from './PhrasesReducer';
+import {seenPhrases} from './SeenPhrasesReducer';
+import {learntPhrases} from './LearntPhrasesReducer';
 import {categoryListId} from './categoryListIdReducer';
 
 export default combineReducers({
