+++ src/screens/LearningScreen/LearningScreen.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import {StyleSheet, SafeAreaView, View, TouchableOpacity} from 'react-native';
-import {leariningScreenManager} from '../../ListManagers/LearningScreenManager';
+import {leariningScreenManager} from '../../listManagers/LearningScreenManager';
 import ScreenHeader from './LeariningScreenHeader';
 import SectionHeading from '../../components/SectionHeading/SectionHeading';
-import PhraseTextarea from '../../components/PhraseTextarea/phraseTextarea';
+import PhraseTextarea from '../../components/PhraseTextarea/PhraseTextarea';
 import ListItem from '../../components/ListItem/ListItem';
 import GlobalStyles from '../../constants/GlobalStyles';
 import NextButton from '../../components/NextButton/NextButton';
